Refactor content creation and detail components: streamline ContentPage and ContentDetailPage for improved content management; enhance CreateContentModal with better validation and loading states; update ContentCard to simplify author display; adjust API types for consistency and clarity.
