+++ src/components/content-detail/PlatformsCard.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import { ExternalLink } from 'lucide-react';
+import { ContentPlatform } from '@/src/types/content';
+
+interface PlatformsCardProps {
+  platforms: ContentPlatform[] | undefined;
+}
+
+export const PlatformsCard: React.FC<PlatformsCardProps> = ({ platforms }) => {
+  if (!platforms?.length) return null;
+
+  return (
+    <div className='bg-white rounded-2xl shadow-sm border border-white/20 p-6'>
+      <h3 className='font-bold text-gray-900 mb-4'>Publishing Platforms</h3>
+      <div className='space-y-3'>
+        {platforms.map((platform: ContentPlatform, index: number) => (
+          <div
+            key={index}
+            className='flex items-center gap-3 p-3 bg-gray-50 rounded-xl hover:bg-gray-100 transition-colors'
+          >
+            <div
+              className={`w-10 h-10 ${platform.color} rounded-lg flex items-center justify-center text-white font-bold`}
+            >
+              {platform.name.charAt(0)}
+            </div>
+            <div className='flex-1'>
+              <div className='font-medium text-gray-900'>{platform.name}</div>
+            </div>
+            <ExternalLink className='w-4 h-4 text-gray-400' />
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+};
