+++ src/components/content-detail/index.ts
@@ -1,0 +1,10 @@
+export { ContentRenderer } from './ContentRenderer';
+export { HeroSection } from './HeroSection';
+export { StatsBar } from './StatsBar';
+export { ContentTabs } from './ContentTabs';
+export { Sidebar } from './Sidebar';
+export { AuthorCard } from './AuthorCard';
+export { Timeline } from './Timeline';
+export { TagsCard } from './TagsCard';
+export { PlatformsCard } from './PlatformsCard';
+export * from './utils';
